Perda de informação

+++ Casting/Casting/Program.cs
@@ -7,9 +7,9 @@ namespace Casting
         static void Main(string[] args)
         {
             double a;
-            float b;
+            int b;
             a = 5.1;
-            b = (float)a;
+            b = (int)a;
             Console.WriteLine(b);
         }
     }
